Add LRUStore datastore wrapper
Package lrustore provides a datastore wrapper that limits the number of items stored to the specified capacity. When at capacity, the oldest items are removed in LRU to make root for new items. Putting and getting items maintains LRU order.

This is useful for implementing a LRU-cache on top of a datastore, or for maintaining limited storage size by deleting items that have not been used for a longer time.