feat: return error from txn.Discard()
To properly discard a transaction usually requires some external call
to the backing DB, which can fail. Currently implementations just
swallow any errors when discarding transactions, but this should be
surfaced to the application so that it can decide the appropriate
action (at the very least, log the error).

This is of course a breaking change and both datastore impls and
consumers will need to be updated.