Results.Close should return error (#230)
Results.Close should always return error. Even if it is not currently used for anything, having the ability to return error allows for higher level functionality to handle error conditiond that may be important for future code. This was removed in v0.8.0, as it did not serve a functional purpose.